# server/openapi.yaml
openapi: 3.1.0
info:
  title: Community Garden API
  version: 1.0.0
  description: |
    Marketplace API connecting local food producers with consumers.
    Supports AI-powered search (FR-08), Future Orders demand signals (FR-11),
    Stripe payments, and SendGrid notifications.
  contact:
    name: Community Garden Team
    url: https://github.com/redwan1795/HACK-TECH-TALK
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.community-garden.app/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRole:
      type: string
      enum: [consumer, producer, broker, admin]
...
    User:
      type: object
      required: [id, email, name, role, licensed, createdAt]
      properties:
        id:          { type: string, format: uuid }
        email:       { type: string, format: email }
        name:        { type: string }
        role:        { $ref: '#/components/schemas/UserRole' }
        locationZip: { type: string }
        licensed:    { type: boolean }
        createdAt:   { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string }
            message: { type: string }
            details: {}
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, role]
              properties:
                email:       { type: string, format: email }
                password:    { type: string, minLength: 8 }
                name:        { type: string }
                role:        { $ref: '#/components/schemas/UserRole' }
                locationZip: { type: string }
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
                  user:         { $ref: '#/components/schemas/User' }
        "409":
          description: Email already in use
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in and receive tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
                  user:         { $ref: '#/components/schemas/User' }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token and issue new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      summary: Invalidate refresh token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "204":
          description: Logged out

  # ── Users ─────────────────────────────────────────────────────────────────────
...